# Faster C code, Rust code