--- a/measure.sh
+++ b/measure.sh
@@ -1,5 +1,20 @@
 #!/bin/bash
 
+# Fordítások
+
+CFILE=proc_speed_test_C
+gcc -O2 -Wall $CFILE.c -o $CFILE
+
+CFILE=proc_speed_test_C2
+gcc -O2 -Wall $CFILE.c -o $CFILE
+
+javac Proc_Speed_Test.java
+
+rustc -O proc_speed_test_Rust.rs
+
+# Fordítások vége
+# Futtatások
+
 # a time format azért ilyen, hogy az '=' kezdetű sorokat közvetlenül .csv-be tudjam menteni,
 # amit aztán egy LibreOffice fel tud dolgozni
 alias tm='/usr/bin/time -f "= \"%C\";%e;%S;%U;%I;%O;%c"'
@@ -25,6 +40,12 @@ tm perl ./proc_speed_test.pl
 
 echo -e "\nC"
 tm ./proc_speed_test_C
+
+echo -e "\nC"
+tm ./proc_speed_test_C2
+
+echo -e "\nC"
+tm ./proc_speed_test_Rust
 
 echo -e "\nLUA"
 tm luajit ./proc_speed_test.lua